moved demo scripts to be in the docs folder, reorganized all markdown files to live in their own directory
