+++ .github/workflows/deploy_docs.yaml
@@ -39,11 +39,11 @@ jobs:
 
       - name: Generate Python docs
         working-directory: ${{runner.workspace}}/fastsim/
-        run: pydoc-markdown -I python/ --render-toc > docs/python-doc.md
+        run: pydoc-markdown -I python/ --render-toc > docs/content/python-doc.md
 
       - name: Convert demos to notebooks
         working-directory: ${{runner.workspace}}/fastsim/
-        run: python docs/demos_to_notebooks.py --dir=python/fastsim/demos
+        run: python docs/demos_to_notebooks.py
 
       - name: Setup Pages
         id: pages
