+++ docs/.gitignore
@@ -2,4 +2,4 @@
 _build
 
 # Generated by deploy-book.yaml workflow
-python-doc.md
+content/python-doc.md
