+++ docs/content/intro.md
@@ -1,0 +1,2 @@
+:::{include} ../../README.md
+:::
