+++ docs/demo_scripts/powertrains/demo_hev.py
