+++ docs/demo_scripts/thermal/demo_hev_thrml_ws_ca.py
@@ -4,20 +4,24 @@ execute:
   skip: true
 ---
 
-# HEV Thermal Demo: Cold Start, Cold Ambient
+# HEV Thermal Demo: Warm Start, Cold Ambient
 
 This demo simulates a hybrid electric vehicle with thermal modeling under
-cold start and cold ambient conditions, where the cabin, battery, and
-engine begin at the same temperature as the surrounding air.
+warm start and cold ambient conditions, where the cabin and battery begin
+warm, the engine begins hot, and the surrounding air is cold.
 """
 
 # %%
 import os
+import sys
+from pathlib import Path
 
 import seaborn as sns
 
+sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
+
 import fastsim as fsim
-from fastsim.demos.plot_utils import (
+from plot_utils import (
     plot_hev_fc_energy,
     plot_hev_fc_pwr,
     plot_hev_res_energy,
@@ -35,24 +39,27 @@ SAVE_FIGS = os.environ.get("SAVE_FIGS", "false").lower() == "true"
 """
 ## Setup and Simulation
 
-Load a thermal HEV, set cold initial temperatures for the cabin, battery,
-and engine to match the cold ambient, and run the simulation.
+Load a thermal HEV, set warm initial temperatures for the cabin and
+battery, a hot initial temperature for the engine, and a cold ambient
+temperature, then run the simulation.
 """
 
 # %%
 celsius_to_kelvin = 273.15
-temp_amb_and_init = -6.7 + celsius_to_kelvin
+temp_amb = -6.7 + celsius_to_kelvin
+temp_init_bat_and_cab = 22.0 + celsius_to_kelvin
+temp_init_eng = 70.0 + celsius_to_kelvin
 
 # load 2021 Hyundai Sonata HEV with thermal model
 veh_dict = fsim.Vehicle.from_resource(
     "2021_Hyundai_Sonata_Hybrid_Blue_thrml.yaml",
 ).to_pydict()
-veh_dict["cabin"]["LumpedCabin"]["state"]["temperature_kelvin"] = temp_amb_and_init
+veh_dict["cabin"]["LumpedCabin"]["state"]["temperature_kelvin"] = temp_init_bat_and_cab
 veh_dict["pt_type"]["HEV"]["res"]["thrml"]["RESLumpedThermal"]["state"]["temperature_kelvin"] = (
-    temp_amb_and_init
+    temp_init_bat_and_cab
 )
 veh_dict["pt_type"]["HEV"]["fc"]["thrml"]["FuelConverterThermal"]["state"]["temperature_kelvin"] = (
-    temp_amb_and_init
+    temp_init_eng
 )
 veh = fsim.Vehicle.from_pydict(veh_dict)
 
@@ -61,7 +68,7 @@ veh.set_save_interval(1)
 # %%
 # load cycle and set ambient temperature
 cyc_dict = fsim.Cycle.from_resource("udds.csv").to_pydict()
-cyc_dict["temp_amb_air_kelvin"] = [temp_amb_and_init] * len(cyc_dict["time_seconds"])
+cyc_dict["temp_amb_air_kelvin"] = [temp_amb] * len(cyc_dict["time_seconds"])
 cyc = fsim.Cycle.from_pydict(cyc_dict)
 
 # %%
@@ -75,7 +82,7 @@ sd_dict = sd.to_pydict(flatten=True)
 ## Visualize Results
 
 Fuel converter power and energy, battery power and energy, component
-temperatures, and road loads under cold start, cold ambient conditions.
+temperatures, and road loads under warm start, cold ambient conditions.
 """
 
 # %%
