--- python/fastsim/demos/.gitignore
@@ -1,3 +1,0 @@
-plots
-*.svg
-*.png
